Fix probability bug
Based on description, under normal circumstances NM should do something 10% of the time. But based on how the code is structured it would be 1/10 for each enabled action. So if you had all four turned on (wifi, response code, response failure, delay) you'd actually see a 40% (4 x 10%) chance of *something* being monkey'd with. Sometimes you'd see multiple things get monkeyed with (a delay *and* a 404).

This fixes that by checking the 1-in-10 probability *first* and then randomly picking one of the enabled actions to perform.